+++ src/OpenTelemetry.Instrumentation.Process/ProcessMetrics.cs
@@ -16,7 +16,7 @@ internal sealed class ProcessMetrics
     internal static readonly AssemblyName AssemblyName = typeof(ProcessMetrics).Assembly.GetName();
     internal static readonly string MeterName = AssemblyName.Name;
 
-    private static readonly Meter MeterInstance = new(MeterName, AssemblyName.Version.ToString());
+    private static readonly Meter MeterInstance = new(MeterName, SignalVersionHelper.GetVersion<ProcessMetrics>());
 
     static ProcessMetrics()
     {