+++ src/OpenTelemetry.Instrumentation.Hangfire/CHANGELOG.md
@@ -5,6 +5,8 @@
 * Update `OpenTelemetry.Api.ProviderBuilderExtensions` to `1.7.0`.
   * Update `OpenTelemetry.Api` to `1.7.0`.
   ([#1508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1508))
+* `ActivitySource.Version` is set to NuGet package version.
+  ([#1624](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1624))
 
 ## 1.6.0-beta.1